+++ rusk-wallet/src/dat.rs
@@ -11,21 +11,26 @@ use std::io::Read;
 
 use wallet_core::Seed;
 
-use crate::crypto::decrypt;
-use crate::{Error, WalletPath};
+use crate::crypto::{decrypt_aes_cbc, decrypt_aes_gcm};
+use crate::{Error, WalletPath, IV_SIZE, SALT_SIZE};
 
 /// Binary prefix for old Dusk wallet files
 pub const OLD_MAGIC: u32 = 0x1d0c15;
 /// Binary prefix for new binary file format
 pub const MAGIC: u32 = 0x72736b;
 /// The latest version of the rusk binary format for wallet dat file
-pub const LATEST_VERSION: Version = (0, 0, 1, 0, false);
+pub const LATEST_VERSION: Version = (0, 0, 2, 0, false);
 /// The type info of the dat file we'll save
 pub const FILE_TYPE: u16 = 0x0200;
 /// Reserved for futures use, 0 for now
 pub const RESERVED: u16 = 0x0000;
 /// (Major, Minor, Patch, Pre, Pre-Higher)
 type Version = (u8, u8, u8, u8, bool);
+
+type Salt = [u8; SALT_SIZE];
+type Iv = [u8; IV_SIZE];
+
+const FILE_HEADER_SIZE: usize = 12;
 
 /// Versions of the potential wallet DAT files we read
 #[derive(Copy, Clone, Debug, PartialEq)]
@@ -42,7 +47,43 @@ impl DatFileVersion {
     /// Checks if the file version is older than the latest Rust Binary file
     /// format
     pub fn is_old(&self) -> bool {
-        matches!(self, Self::Legacy | Self::OldWalletCli(_))
+        match self {
+            Self::Legacy | Self::OldWalletCli(_) => true,
+            Self::RuskBinaryFileFormat(version) => version < &LATEST_VERSION,
+        }
+    }
+}
+
+fn read_salt_and_iv(
+    version: DatFileVersion,
+    bytes: &[u8],
+) -> Result<Option<(Salt, Iv)>, Error> {
+    match version {
+        DatFileVersion::Legacy | DatFileVersion::OldWalletCli(_) => Ok(None),
+        DatFileVersion::RuskBinaryFileFormat(version)
+            if version_without_pre_higher(version) < (0, 0, 2, 0) =>
+        {
+            Ok(None)
+        }
+        DatFileVersion::RuskBinaryFileFormat(_) => {
+            if let (Some(salt_bytes), Some(iv_bytes)) = (
+                bytes.get(FILE_HEADER_SIZE..FILE_HEADER_SIZE + SALT_SIZE),
+                bytes.get(
+                    FILE_HEADER_SIZE + SALT_SIZE
+                        ..FILE_HEADER_SIZE + SALT_SIZE + IV_SIZE,
+                ),
+            ) {
+                let salt = salt_bytes
+                    .try_into()
+                    .map_err(|_| Error::WalletFileCorrupted)?;
+                let iv = iv_bytes
+                    .try_into()
+                    .map_err(|_| Error::WalletFileCorrupted)?;
+                Ok(Some((salt, iv)))
+            } else {
+                Err(Error::WalletFileCorrupted)
+            }
+        }
     }
 }
 
@@ -50,7 +91,8 @@ impl DatFileVersion {
 pub(crate) fn get_seed_and_address(
     file: DatFileVersion,
     mut bytes: Vec<u8>,
-    pwd: &[u8],
+    aes_key: &[u8],
+    iv: Option<&[u8; IV_SIZE]>,
 ) -> Result<(Seed, u8), Error> {
     match file {
         DatFileVersion::Legacy => {
@@ -58,7 +100,7 @@ pub(crate) fn get_seed_and_address(
                 bytes.drain(..3);
             }
 
-            bytes = decrypt(&bytes, pwd)?;
+            bytes = decrypt_aes_cbc(&bytes, aes_key)?;
 
             // get our seed
             let seed = bytes[..]
@@ -72,7 +114,7 @@ pub(crate) fn get_seed_and_address(
 
             let result: Result<(Seed, u8), Error> = match (major, minor) {
                 (1, 0) => {
-                    let content = decrypt(&bytes, pwd)?;
+                    let content = decrypt_aes_cbc(&bytes, aes_key)?;
                     let buff = &content[..];
 
                     let seed = buff
@@ -82,7 +124,7 @@ pub(crate) fn get_seed_and_address(
                     Ok((seed, 1))
                 }
                 (2, 0) => {
-                    let content = decrypt(&bytes, pwd)?;
+                    let content = decrypt_aes_cbc(&bytes, aes_key)?;
                     let buff = &content[..];
 
                     // extract seed
@@ -98,10 +140,26 @@ pub(crate) fn get_seed_and_address(
 
             result
         }
-        DatFileVersion::RuskBinaryFileFormat(_) => {
-            let rest = bytes.get(12..(12 + 96));
+        DatFileVersion::RuskBinaryFileFormat(version) => {
+            const OLD_PAYLOAD_SIZE: usize = 96;
+            const PAYLOAD_SIZE: usize = 81;
+
+            let (rest, use_aes_gcm) =
+                if version_without_pre_higher(version) < (0, 0, 2, 0) {
+                    let offset = FILE_HEADER_SIZE;
+                    (bytes.get(offset..(offset + OLD_PAYLOAD_SIZE)), false)
+                } else {
+                    let offset = FILE_HEADER_SIZE + SALT_SIZE + IV_SIZE;
+                    (bytes.get(offset..(offset + PAYLOAD_SIZE)), true)
+                };
+
             if let Some(rest) = rest {
-                let content = decrypt(rest, pwd)?;
+                let content = if use_aes_gcm {
+                    let iv = iv.ok_or(Error::WalletFileCorrupted)?;
+                    decrypt_aes_gcm(rest, aes_key, iv)?
+                } else {
+                    decrypt_aes_cbc(rest, aes_key)?
+                };
 
                 if let Some(seed_buff) = content.get(0..65) {
                     let seed = seed_buff[0..64]
@@ -210,9 +268,34 @@ pub fn read_file_version(file: &WalletPath) -> Result<DatFileVersion, Error> {
     check_version(Some(&header_buf))
 }
 
+/// Read the file version of the dat file from the header and, if present,
+/// the salt and IV.
+pub fn read_file_version_and_salt_iv(
+    file: &WalletPath,
+) -> Result<(DatFileVersion, Option<(Salt, Iv)>), Error> {
+    let path = &file.wallet;
+
+    if !path.is_file() {
+        return Err(Error::WalletFileMissing);
+    }
+
+    let mut fs = fs::File::open(path)?;
+    let mut buf = [0; FILE_HEADER_SIZE + SALT_SIZE + IV_SIZE];
+    fs.read_exact(&mut buf)?;
+    let version = check_version(Some(&buf[..FILE_HEADER_SIZE]))?;
+    let salt = read_salt_and_iv(version, &buf)?;
+
+    Ok((version, salt))
+}
+
 pub(crate) fn version_bytes(version: Version) -> [u8; 4] {
     u32::from_be_bytes([version.0, version.1, version.2, version.3])
         .to_be_bytes()
+}
+
+/// Returns the given version with its last item, pre-higher, removed
+pub fn version_without_pre_higher(version: Version) -> (u8, u8, u8, u8) {
+    (version.0, version.1, version.2, version.3)
 }
 
 #[cfg(test)]
@@ -227,9 +310,13 @@ mod tests {
         let legacy_file = vec![
             0xab, 0x38, 0x81, 0x3b, 0xfc, 0x79, 0x11, 0xf9, 0x86, 0xd6, 0xd0,
         ];
-        // new header
-        let new_file = vec![
+        // rusk binary headers
+        let rusk_bin_file_1 = vec![
             0x00, 0x72, 0x73, 0x6b, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
+            0x00,
+        ];
+        let rusk_bin_file_2 = vec![
+            0x00, 0x72, 0x73, 0x6b, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
             0x00,
         ];
 
@@ -244,8 +331,13 @@ mod tests {
         );
 
         assert_eq!(
-            check_version(Some(&new_file)).unwrap(),
+            check_version(Some(&rusk_bin_file_1)).unwrap(),
             DatFileVersion::RuskBinaryFileFormat((0, 0, 1, 0, false))
+        );
+
+        assert_eq!(
+            check_version(Some(&rusk_bin_file_2)).unwrap(),
+            DatFileVersion::RuskBinaryFileFormat((0, 0, 2, 0, false))
         );
     }
 }